part of UGENE-1410 with msvc 2013 support (2)
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@7688 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Gui/src/util/imageExport/WidgetScreenshotExportTask.cpp
@@ -68,7 +68,7 @@ void WidgetScreenshotExportToSvgTask::run() {
 
     file.close();
     QDomNodeList radialGradients=doc.elementsByTagName("radialGradient");
-    for (int i = 0;i < radialGradients.length(); i++) {
+    for (int i = 0;i < static_cast<int>(radialGradients.length()); i++) {
         if (radialGradients.at(i).isElement()) {
             QDomElement tag = radialGradients.at(i).toElement();
             if (tag.hasAttribute("xml:id")) {
